+++ bedrock/products/templates/products/vpn/landing.html
@@ -33,12 +33,6 @@
 {% set _utm_campaign = 'vpn-product-page' %}
 {% set _params = '?utm_source=' ~ _utm_source ~ '&utm_medium=referral&utm_campaign=' ~ _utm_campaign %}
 
-{% block experiments %}
-  {% if switch('experiment-vpn-pricing-position') %}
-    {{ js_bundle('mozilla-vpn-landing-pricing-position-experiment') }}
-  {% endif %}
-{% endblock %}
-
 {% block page_css %}
   {{ css_bundle('mozilla-vpn-landing') }}
 {% endblock %}
@@ -96,7 +90,39 @@
       <p>{{ ftl('vpn-landing-privacy-desc') }}</p>
     {% endcall %}
 
-    {% block top_pricing_position %}{% endblock %}
+    {# note: `id=#pricing` is used as an anchor link from android in-product subscription flows, so do not remove (issue 10039) #}
+    <div id="pricing" class="mzp-is-anchor-link">
+      {% if vpn_available %}
+        {% include 'products/vpn/includes/pricing-plus-relay.html' %}
+      {% else %}
+        <div class="vpn-waitlist-feature-block">
+          {% call vpn_content_block(
+            class_name='vpn-content-block-price t-highlight'
+          ) %}
+          <div class="l-columns-two">
+            <div class="l-column-first">
+              <a class="mzp-c-button mzp-t-xl" href="{{ url('products.vpn.invite') }}" data-cta-type="button" data-cta-text="Join the VPN Waitlist" data-cta-position="secondary">
+                {{ ftl('vpn-shared-waitlist-link') }}
+              </a>
+
+              <p class="availability-copy">
+                {{ ftl('vpn-shared-available-countries-v6') }}
+              </p>
+            </div>
+            <div class="l-column-last">
+              <ul class="mzp-u-list-styled vpn-feature-list">
+                <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-devices', devices=connect_devices) }}</li>
+                <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-servers', servers=connect_servers, countries=connect_countries) }}</li>
+                <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-encrypt') }}</li>
+                <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-bandwidth') }}</li>
+                <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-activity') }}</li>
+              </ul>
+            </div>
+          </div>
+          {% endcall %}
+        </div>
+      {% endif %}
+    </div>
 
     {% call vpn_content_media(
       heading=ftl('vpn-landing-fast-secure-heading'),
@@ -217,42 +243,6 @@
         {% endcall %}
       </div>
     {% endif %}
-
-    {% block default_pricing_position %}
-      {# note: `id=#pricing` is used as an anchor link from android in-product subscription flows, so do not remove (issue 10039) #}
-      <div id="pricing" class="mzp-is-anchor-link">
-        {% if vpn_available %}
-          {% include 'products/vpn/includes/pricing-plus-relay.html' %}
-        {% else %}
-          <div class="vpn-waitlist-feature-block">
-            {% call vpn_content_block(
-              class_name='vpn-content-block-price t-highlight'
-            ) %}
-            <div class="l-columns-two">
-              <div class="l-column-first">
-                <a class="mzp-c-button mzp-t-xl" href="{{ url('products.vpn.invite') }}" data-cta-type="button" data-cta-text="Join the VPN Waitlist" data-cta-position="secondary">
-                  {{ ftl('vpn-shared-waitlist-link') }}
-                </a>
-
-                <p class="availability-copy">
-                  {{ ftl('vpn-shared-available-countries-v6') }}
-                </p>
-              </div>
-              <div class="l-column-last">
-                <ul class="mzp-u-list-styled vpn-feature-list">
-                  <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-devices', devices=connect_devices) }}</li>
-                  <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-servers', servers=connect_servers, countries=connect_countries) }}</li>
-                  <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-encrypt') }}</li>
-                  <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-bandwidth') }}</li>
-                  <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-activity') }}</li>
-                </ul>
-              </div>
-            </div>
-            {% endcall %}
-          </div>
-        {% endif %}
-      </div>
-    {% endblock %}
   </div>
 
   <section id="faq" class="vpn-faq mzp-l-content mzp-t-content-lg">
