--- bedrock/products/templates/products/vpn/variants/landing-2.html
@@ -1,45 +1,0 @@
-{#
- This Source Code Form is subject to the terms of the Mozilla Public
- License, v. 2.0. If a copy of the MPL was not distributed with this
- file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
-{% extends "products/vpn/landing.html" %}
-
-{% block top_pricing_position %}
-  {# note: `id=#pricing` is used as an anchor link from android in-product subscription flows, so do not remove (issue 10039) #}
-  <div id="pricing" class="mzp-is-anchor-link c-pricing-top">
-    {% if vpn_available %}
-      {% include 'products/vpn/includes/pricing-plus-relay.html' %}
-    {% else %}
-      <div class="vpn-waitlist-feature-block">
-        {% call vpn_content_block(
-          class_name='vpn-content-block-price t-highlight'
-        ) %}
-        <div class="l-columns-two">
-          <div class="l-column-first">
-            <a class="mzp-c-button mzp-t-xl" href="{{ url('products.vpn.invite') }}" data-cta-type="button" data-cta-text="Join the VPN Waitlist" data-cta-position="secondary">
-              {{ ftl('vpn-shared-waitlist-link') }}
-            </a>
-
-            <p class="availability-copy">
-              {{ ftl('vpn-shared-available-countries-v6') }}
-            </p>
-          </div>
-          <div class="l-column-last">
-            <ul class="mzp-u-list-styled vpn-feature-list">
-              <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-devices', devices=connect_devices) }}</li>
-              <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-servers', servers=connect_servers, countries=connect_countries) }}</li>
-              <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-encrypt') }}</li>
-              <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-bandwidth') }}</li>
-              <li class="vpn-feature-list-item">{{ ftl('vpn-shared-features-activity') }}</li>
-            </ul>
-          </div>
-        </div>
-        {% endcall %}
-      </div>
-    {% endif %}
-  </div>
-{% endblock %}
-
-{% block default_pricing_position %}{% endblock %}
