--- bedrock/products/templates/products/vpn/variants/landing-refresh-2.html
@@ -1,31 +1,0 @@
-{#
- This Source Code Form is subject to the terms of the Mozilla Public
- License, v. 2.0. If a copy of the MPL was not distributed with this
- file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
-{% extends "products/vpn/landing-refresh.html" %}
-
-{% block top_pricing_position %}
-{# note: `id=#pricing` is used as an anchor link from android in-product subscription flows, so do not remove (issue 10039) #}
-<section id="pricing" class="mzp-l-content mzp-t-content-xl">
-  {% if vpn_available %}
-    <header class="c-pricing-main-header">
-      <h3 class="c-section-title">{{ ftl('vpn-landing-one-subscription-for-all-your') }}</h3>
-    </header>
-
-    {% include 'products/vpn/includes/pricing-refresh.html' %}
-
-  {% else %}
-    <header class="c-pricing-main-header">
-      <h2 class="c-section-title">{{ ftl('vpn-shared-mozilla-vpn-is-not-yet-available') }}</h2>
-
-      <a class="mzp-c-button mzp-t-product mzp-t-xl" href="{{ url('products.vpn.invite') }}" data-cta-type="button" data-cta-text="Join the VPN Waitlist">
-        {{ ftl('vpn-shared-waitlist-link') }}
-      </a>
-    </header>
-  {% endif %}
-</section>
-{% endblock %}
-
-{% block default_pricing_position %}{% endblock %}
