+++ media/css/products/vpn/common-refresh.scss
@@ -14,6 +14,15 @@ main {
         line-height: $vpn-title-line-height;
     }
 
+    .u-title-xl {
+        @include font-size($vpn-title-lg-size);
+        line-height: 1.2;
+
+        @media #{$mq-md} {
+            @include font-size($vpn-title-xl-size);
+        }
+    }
+
     .u-title-lg {
         @include font-size($vpn-title-md-size);
         line-height: 1.2;
@@ -32,6 +41,15 @@ main {
         }
     }
 
+    .u-title-xs {
+        @include font-size($vpn-title-2xs-size);
+        line-height: 1.2;
+
+        @media #{$mq-md} {
+            @include font-size($vpn-title-xs-size);
+        }
+    }
+
     .u-body-lg {
         @include text-body-lg;
     }
@@ -44,6 +62,11 @@ main {
     @media #{$mq-md} {
         @include font-size($vpn-title-2xl-size);
     }
+}
+
+.c-page-subtitle {
+    font-weight: normal;
+    margin-bottom: $spacing-2xl;
 }
 
 .c-section-title {
